partial pr subscription operation PR
