+++ locksmith/src/controllers/v2/subscriptionController.ts
@@ -1,37 +1,9 @@
 import { Response, Request } from 'express'
 import { KeySubscription } from '../../models'
 import normalizer from '../../utils/normalizer'
-import {
-  SubgraphService,
-  Web3Service,
-  getErc20BalanceForAddress,
-  getErc20Decimals,
-  getAllowance,
-  getErc20TokenSymbol,
-} from '@unlock-protocol/unlock-js'
-import networks from '@unlock-protocol/networks'
-import { ethers } from 'ethers'
-import { Op } from 'sequelize'
-import dayjs from 'dayjs'
-import relative from 'dayjs/plugin/relativeTime'
-import duration from 'dayjs/plugin/duration'
-
-dayjs.extend(relative)
-dayjs.extend(duration)
-
-interface Amount {
-  amount: string
-  decimals: number
-  symbol: string
-}
-export interface Subscription {
-  next: number | null
-  balance: Amount
-  price: Amount
-  possibleRenewals: string
-  approvedRenewals: string
-  type: 'Crypto' | 'Stripe'
-}
+import subscriptionOperations, {
+  Subscription,
+} from '../../operations/subscriptionOperations'
 
 export class SubscriptionController {
   /**
@@ -42,117 +14,18 @@ export class SubscriptionController {
     const lockAddress = normalizer.ethereumAddress(request.params.lockAddress)
     const keyId = Number(request.params.keyId)
     const userAddress = normalizer.ethereumAddress(request.user!.walletAddress)
-    const subgraphService = new SubgraphService(networks)
 
-    const key = await subgraphService.key(
-      {
-        where: {
-          tokenId: keyId.toString(),
-          lock: lockAddress.toLowerCase(),
-          owner: userAddress.toLowerCase(),
-        },
-      },
-      {
+    const subscriptions: Subscription[] =
+      await subscriptionOperations.getSubscriptionsForLockByOwner({
+        tokenId: keyId.toString(),
+        owner: userAddress,
         network,
-      }
-    )
-
-    // If no key is found or not erc20 or version < 11 which we don't fully support, return nothing.
-    if (
-      !key ||
-      key.lock.tokenAddress === ethers.constants.AddressZero ||
-      parseInt(key.lock.version) < 11
-    ) {
-      return response.status(200).send({
-        subscriptions: [],
-      })
-    }
-
-    const web3Service = new Web3Service(networks)
-    const provider = web3Service.providerForNetwork(network)
-    const [userBalance, decimals, userAllowance, symbol] = await Promise.all([
-      getErc20BalanceForAddress(key.lock.tokenAddress, userAddress, provider),
-      getErc20Decimals(key.lock.tokenAddress, provider),
-      getAllowance(
-        key.lock.tokenAddress,
-        key.lock.address,
-        provider,
-        userAddress
-      ),
-      getErc20TokenSymbol(key.lock.tokenAddress, provider),
-    ])
-
-    const balance = ethers.utils.formatUnits(userBalance, decimals)
-
-    const price = ethers.BigNumber.from(key.lock.price)
-    const next =
-      key.expiration === ethers.constants.MaxUint256.toString()
-        ? null
-        : dayjs.unix(key.expiration).isBefore(dayjs())
-        ? null
-        : parseInt(key.expiration)
-
-    // Approved renewals
-    const approvedRenewalsAmount =
-      userAllowance.gt(0) && price.gt(0)
-        ? userAllowance.div(price)
-        : ethers.BigNumber.from(0)
-
-    const approvedRenewals = approvedRenewalsAmount.toString()
-
-    const info = {
-      next,
-      balance: {
-        amount: balance,
-        decimals,
-        symbol,
-      },
-      price: {
-        amount: ethers.utils.formatUnits(price, decimals),
-        decimals,
-        symbol,
-      },
-    }
-
-    const stripeSubscription = await KeySubscription.findOne({
-      where: {
-        keyId,
         lockAddress,
-        network,
-        userAddress,
-        recurring: {
-          [Op.gt]: 0,
-        },
-      },
-    })
-
-    const subscriptions: Subscription[] = []
-
-    // if card subscription is found, add it.
-    if (stripeSubscription) {
-      const approvedRenewals = stripeSubscription.recurring?.toString()
-
-      subscriptions.push({
-        ...info,
-        approvedRenewals,
-        possibleRenewals: approvedRenewals,
-        type: 'Stripe',
       })
-    }
 
-    // Add the default crypto subscription details.
-    const cryptoSubscription: Subscription = {
-      ...info,
-      approvedRenewals: approvedRenewals,
-      possibleRenewals: ethers.BigNumber.from(userBalance)
-        .div(price)
-        .toString(),
-      type: 'Crypto',
-    }
-
-    subscriptions.push(cryptoSubscription)
-
-    return response.status(200).send({ subscriptions })
+    return response.status(200).send({
+      subscriptions,
+    })
   }
 
   /**
